mediaAnalyser app: split in several files
Rename AvSoudFiles to AvSoudFile.
